modify game response and enable notification deletion

+++ src/main/java/com/tejko/yamb/business/services/NotificationServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import com.tejko.yamb.business.interfaces.NotificationService;
@@ -15,6 +16,7 @@ import com.tejko.yamb.domain.models.Notification;
 import com.tejko.yamb.domain.models.Player;
 import com.tejko.yamb.domain.repositories.NotificationRepository;
 import com.tejko.yamb.domain.repositories.PlayerRepository;
+import com.tejko.yamb.security.AuthContext;
 
 @Service
 public class NotificationServiceImpl implements NotificationService {
@@ -55,12 +57,20 @@ public class NotificationServiceImpl implements NotificationService {
     @Override
     public void deleteByExternalId(UUID externalId) {
         Notification notification = getByExternalId(externalId);
+        checkPermission(notification.getPlayer().getExternalId());
         notificationRepo.delete(notification);
     }
     
     @Override
     public void deleteAll() {
         notificationRepo.deleteAll();
+    }
+
+    private void checkPermission(UUID playerExternalId) {
+        Player authenticatedPlayer = AuthContext.getAuthenticatedPlayer();  
+        if (playerExternalId == null || authenticatedPlayer != null && !authenticatedPlayer.getExternalId().equals(playerExternalId)) {
+            throw new AccessDeniedException("error.access_denied");
+        }
     }
     
 }
